# src/test/resources/Collection_Service_API_Spec_v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "https://<host>:<port>/cp/collection-service/",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/v1/configurations": {
            "get": {
                "tags": [
                    "configuration-controller"
                ],
                "summary": "Export the configurations",
                "operationId": "exportConfig",
                "responses": {
                    "200": {
                        "description": "Configurations exported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AllConfigurations"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error exporting the configurations."
                    }
                }
            },
            "post": {
                "tags": [
                    "configuration-controller"
                ],
                "summary": "Import the configurations",
                "operationId": "importConfig",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AllConfigurations"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Configurations imported successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AllConfigurations"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error importing the configurations."
                    }
                }
            },
            "delete": {
                "tags": [
                    "configuration-controller"
                ],
                "summary": "Clear the collector configurations",
                "operationId": "deleteConfig",
                "responses": {
                    "204": {
                        "description": "All configurations are deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMessage"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error deleting the configurations."
                    }
                }
            }
        },
        "/api/v1/configurations/import/wae": {
            "post": {
                "tags": [
                    "configuration-controller"
                ],
                "summary": "Import the configurations from WAE",
                "operationId": "importFromWae",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Below Configurations are imported successfully : Auth groups,SNMP groups,network access,node filter configurations,agent configurations, network configurations part of the composer workflow, schedulers configurations and the Archive configurations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseMessage"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error importing the configurations."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "SnmpGroupData": {
                "required": [
                    "name",
                    "snmpType"
                ],
                "type": "object",
                "properties": {
                    "snmpType": {
                        "type": "string",
                        "enum": [
                            "SNMPv2c",
                            "SNMPv3"
                        ]
                    },
                    "securityLevel": {
                        "type": "string",
                        "enum": [
                            "AUTH_PRIV",
                            "NOAUTH_PRIV",
                            "NOAUTH_NOPRIV"
                        ]
                    },
                    "username": {
                        "type": "string"
                    },
                    "authenticationProtocol": {
                        "type": "string",
                        "enum": [
                            "SHA",
                            "MD5"
                        ]
                    },
                    "authenticationPassword": {
                        "type": "string"
                    },
                    "encryptionProtocol": {
                        "type": "string",
                        "enum": [
                            "AES",
                            "DES"
                        ]
                    },
                    "encryptionPassword": {
                        "type": "string"
                    },
                    "roCommunity": {
                        "type": "string"
                    },
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "SchedulerConfigData": {
                "required": [
                    "name",
                    "networkId"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "networkId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "networkName": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "cronExpression": {
                        "type": "string"
                    },
                    "runNow": {
                        "type": "boolean"
                    },
                    "taskConfigDataList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskConfigData"
                        }
                    },
                    "active": {
                        "type": "boolean"
                    }
                }
            },
            "TaskConfigData": {
                "type": "object",
                "properties": {
                    "taskName": {
                        "type": "string"
                    },
                    "collectorId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "collectorName": {
                        "type": "string"
                    },
                    "collect": {
                        "type": "boolean"
                    },
                    "aggreagate": {
                        "type": "boolean"
                    },
                    "archive": {
                        "type": "boolean"
                    }
                }
            },
...
            "NodeFilterData": {
                "required": [
                    "condition",
                    "type",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "HOST_INDIVIDUAL",
                            "HOST_REGEX",
                            "IP_CIDR",
                            "IP_REGEX",
                            "IP_INDIVIDUAL"
                        ]
                    },
                    "value": {
                        "type": "string"
                    },
                    "condition": {
                        "type": "string",
                        "enum": [
                            "INCLUDE",
                            "EXCLUDE"
                        ]
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "NodeListData": {
                "required": [
                    "nodeIp"
                ],
                "type": "object",
                "properties": {
                    "nodeIp": {
                        "type": "string"
                    },
                    "nodeManagementIp": {
                        "type": "string"
                    },
                    "authGroupName": {
                        "type": "string"
                    },
                    "snmpGroupName": {
                        "type": "string"
                    },
                    "snmpGroupMismatch": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "authGroupMismatch": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "NodeProfileData": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "defaultAuthGroup": {
                        "type": "string"
                    },
                    "defaultSnmpGroup": {
                        "type": "string"
                    },
                    "useNodeListAsIncludeFilter": {
                        "type": "boolean"
                    },
                    "nodeList": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NodeListData"
                        }
                    },
                    "nodeFilter": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NodeFilterData"
                        }
                    }
                }
            },
...
            "AllNodeProfileData": {
                "type": "object",
                "properties": {
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "useNodeListAsIncludeFilter": {
                        "type": "boolean"
                    }
                }
            },
...
            "CollectorDataView": {
                "required": [
                    "name",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "TOPO_IGP",
                            "TOPO_BGPLS_XTC",
                            "TOPO_BGP",
                            "LSP_SNMP",
                            "LSP_PCEP_XTC",
                            "CONFIG_PARSE",
                            "TOPO_VPN",
                            "TRAFFIC_POLL",
                            "INVENTORY",
                            "TRAFFIC_DEMAND",
                            "LOGIN_FIND_MULTICAST",
                            "LOGIN_POLL_MULTICAST",
                            "SNMP_FIND_MULTICAST",
                            "SNMP_POLL_MULTICAST",
                            "NETFLOW",
                            "LAYOUT",
                            "EXTERNAL_SCRIPT",
                            "DARE",
                            "MULTICAST",
                            "DEMAND_MESH_CREATOR",
                            "DEMAND_FOR_LSPS",
                            "DEMAND_FOR_P2MP_LSPS",
                            "DEMAND_DEDUCTION",
                            "COPY_DEMANDS"
                        ]
                    },
                    "consolidationType": {
                        "type": "string",
                        "enum": [
                            "DARE",
                            "SAGE",
                            "NONE"
                        ]
                    },
                    "params": {
                        "type": "string"
                    },
                    "sourceCollector": {
                        "$ref": "#/components/schemas/SourceCollector"
                    },
                    "agents": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AllAgentData"
                        }
                    }
                }
            },
...
            "NetworkDataView": {
                "required": [
                    "name",
                    "nodeProfileData"
                ],
                "type": "object",
                "properties": {
                    "nodeProfileData": {
                        "$ref": "#/components/schemas/AllNodeProfileData"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "draft": {
                        "type": "boolean"
                    },
                    "draftConfig": {
                        "type": "string"
                    },
                    "collectors": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CollectorDataView"
                        }
                    }
                }
            },
            "SourceCollector": {
                "required": [
                    "id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "TOPO_IGP",
                            "TOPO_BGPLS_XTC",
                            "TOPO_BGP",
                            "LSP_SNMP",
                            "LSP_PCEP_XTC",
                            "CONFIG_PARSE",
                            "TOPO_VPN",
                            "TRAFFIC_POLL",
                            "INVENTORY",
                            "TRAFFIC_DEMAND",
                            "LOGIN_FIND_MULTICAST",
                            "LOGIN_POLL_MULTICAST",
                            "SNMP_FIND_MULTICAST",
                            "SNMP_POLL_MULTICAST",
                            "NETFLOW",
                            "LAYOUT",
                            "EXTERNAL_SCRIPT",
                            "DARE",
                            "MULTICAST",
                            "DEMAND_MESH_CREATOR",
                            "DEMAND_FOR_LSPS",
                            "DEMAND_FOR_P2MP_LSPS",
                            "DEMAND_DEDUCTION",
                            "COPY_DEMANDS"
                        ]
                    }
                }
            },
...
            "AgentData": {
                "required": [
                    "name",
                    "params",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "SR_PCE_AGENT",
                            "NETFLOW_AGENT",
                            "TRAFFIC_POLLER"
                        ]
                    },
                    "params": {
                        "type": "string"
                    }
                }
            },
            "AuthGroupData": {
                "required": [
                    "name"
                ],
                "type": "object",
                "properties": {
                    "updateDate": {
                        "type": "string",
                        "format": "date",
                        "readOnly": true
                    },
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "loginType": {
                        "type": "string",
                        "enum": [
                            "TELNET",
                            "SSH"
                        ]
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "confirmPassword": {
                        "type": "string"
                    }
                }
            },
            "AllConfigurations": {
                "type": "object",
                "properties": {
                    "snmpGroupDataList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SnmpGroupData"
                        }
                    },
                    "authGroupDataList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuthGroupData"
                        }
                    },
                    "nodeProfileDataList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NodeProfileData"
                        }
                    },
                    "agentDataList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AgentData"
                        }
                    },
                    "networkDataList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NetworkDataView"
                        }
                    },
                    "schedulerConfigDataList": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SchedulerConfigData"
                        }
                    }
                }
            },
            "ResponseMessage": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}